Refactor whereis function to return file path, filename and extension.
The whereis function has been refactored to return the file path, filename and extension of a given input. The code now checks if the input is a character array before proceeding with the search. If there are multiple matches, it prints all of them along with an indication if they are shadowed or not. If there is only one match, it prints just that one. Finally, it returns the path, filename and extension as separate outputs.