Add smooth2a function for 2D array data smoothing
This commit adds the smooth2a function, which can be used to smooth 2D array data. The function applies a mean filter over a rectangle of size (2*Nr+1)-by-(2*Nc+1), replacing each element with the mean of the rectangle centered on it. NaN elements are ignored in the averaging process. If an element is NaN, it will remain NaN in the output. At the edges of the matrix, where a full rectangle cannot be built, as much of the fitting rectangle as possible is used. The function takes three arguments: "matrixIn" (the original matrix), "Nr" (the number of points used to smooth rows), and "Nc" (the number of points used to smooth columns). If Nc is not specified, it defaults to Nr. The output is a smoothed version of the original matrix.

Written by Greg Reeves in March 2009, inspired by Kelly Hilands' "smooth2" from October 2004 and Olof Liungman's code from 1997.

Initial error statements and definitions are included at the beginning of the function. Matrices eL and eR are created to compute running sums for row and column smoothing respectively. NaN elements in matrixIn are set to zero so they don't affect summation calculations. A normalization factor is computed using matrices eL and eR to count non-NaN elements for each element in order to calculate means accurately. Finally, means are calculated using eL, matrixIn, and eR matrices.

Closes #123